+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientTest.kt
@@ -30,6 +30,7 @@ import at.asitplus.wallet.lib.agent.ValidatorSdJwt
 import at.asitplus.wallet.lib.agent.Verifier.VerifyCredentialResult
 import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.*
+import at.asitplus.wallet.lib.data.MediaTypes
 import at.asitplus.wallet.lib.data.rfc3986.toUri
 import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.extensions.supportedSdAlgorithms
@@ -255,17 +256,25 @@ val OpenId4VciClientTest by testSuite {
                     val requestBody = request.body.toByteArray().decodeToString()
                     val authn = request.headers[HttpHeaders.Authorization].shouldNotBeNull()
                     val params = vckJsonSerializer.decodeFromString<CredentialRequestParameters>(requestBody)
+                    // TODO Test with encryption
                     credentialIssuer.credential(
                         authorizationHeader = authn,
-                        params = params,
+                        params = WalletService.CredentialRequest.Plain(params),
                         credentialDataProvider = credentialDataProvider,
                         request = request.toRequestInfo(),
                     ).fold(
                         onSuccess = {
-                            respond(
-                                vckJsonSerializer.encodeToString<CredentialResponseParameters>(it),
-                                headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
-                            )
+                            when (it) {
+                                is CredentialIssuer.CredentialResponse.Encrypted -> respond(
+                                    it.response.serialize(),
+                                    headers = headersOf(HttpHeaders.ContentType, MediaTypes.Application.JWT)
+                                )
+
+                                is CredentialIssuer.CredentialResponse.Plain -> respond(
+                                    vckJsonSerializer.encodeToString<CredentialResponseParameters>(it.response),
+                                    headers = headersOf(HttpHeaders.ContentType, MediaTypes.Application.JSON)
+                                )
+                            }
                         },
                         onFailure = { respondOAuth2Error(it) }
                     )