+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/WalletService.kt
@@ -27,6 +27,7 @@ import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.jws.SdJwtSigned
 import at.asitplus.wallet.lib.jws.SignJwt
 import at.asitplus.wallet.lib.oauth2.OAuth2Client
+import at.asitplus.wallet.lib.oidvci.CredentialIssuer.CredentialResponse
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception.InvalidRequest
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception.InvalidToken
 import com.benasher44.uuid.uuid4
@@ -79,6 +80,13 @@ class WalletService(
          * type `application/jwt` (see [at.asitplus.wallet.lib.data.MediaTypes.Application.JWT]).
          */
         data class Encrypted(val request: JweEncrypted) : CredentialRequest
+
+        fun parse(input: String): KmmResult<CredentialRequest> = catching {
+            if (input.count { it == '.' } == 4)
+                Encrypted(JweEncrypted.deserialize(input).getOrThrow())
+            else
+                Plain(joseCompliantSerializer.decodeFromString<CredentialRequestParameters>(input))
+        }
     }
 
     data class RequestOptions(
@@ -173,7 +181,7 @@ class WalletService(
      * Also send along the [TokenResponseParameters.accessToken] from the token response in HTTP header `Authorization`
      * see [TokenResponseParameters.toHttpHeaderValue].
      * Be sure to include a DPoP header if [TokenResponseParameters.tokenType] is `DPoP`,
-     * see [at.asitplus.wallet.lib.oidvci.BuildDPoPHeader].
+     * see [BuildDPoPHeader].
      * For sample ktor code see `OpenId4VciClient` in `vck-openid-ktor`.
      *
      * @param tokenResponse from the authorization server token endpoint
@@ -251,13 +259,41 @@ class WalletService(
      * decrypting the response if required.
      */
     public suspend fun parseCredentialResponse(
-        response: CredentialResponseParameters,
+        response: String,
+        isEncrypted: Boolean,
         representation: CredentialRepresentation,
         scheme: ConstantIndex.CredentialScheme,
     ): KmmResult<Collection<Holder.StoreCredentialInput>> = catching {
-        response.extractCredentials()
+        response.decryptResponse(isEncrypted)
+            .extractCredentials()
             .map { encryptionService.decrypt(it).getOrThrow() }
             .map { it.toStoreCredentialInput(representation, scheme) }
+    }
+
+    private suspend fun String.decryptResponse(encrypted: Boolean) =
+        vckJsonSerializer.decodeFromString<CredentialResponseParameters>(
+            if (encrypted) encryptionService.decrypt(this).getOrThrow() else this
+        )
+
+    /**
+     * Parses [response] received from the credential issuer, mapping to [Holder.StoreCredentialInput],
+     * decrypting the response if required.
+     */
+    public suspend fun parseCredentialResponse(
+        response: CredentialResponse,
+        representation: CredentialRepresentation,
+        scheme: ConstantIndex.CredentialScheme,
+    ): KmmResult<Collection<Holder.StoreCredentialInput>> = catching {
+        response.decryptResponse()
+            .extractCredentials()
+            .map { encryptionService.decrypt(it).getOrThrow() }
+            .map { it.toStoreCredentialInput(representation, scheme) }
+    }
+
+    private suspend fun CredentialResponse.decryptResponse() = when (this) {
+        is CredentialResponse.Plain -> response
+        is CredentialResponse.Encrypted -> encryptionService.decrypt(response.serialize()).getOrThrow()
+            .decryptResponse(true)
     }
 
     private fun Set<AuthorizationDetails>.toCredentialRequest(): List<CredentialRequestParameters> =