+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciCodeFlowTest.kt
@@ -167,7 +167,10 @@ val OidvciCodeFlowTest by testSuite {
             ).getOrThrow().first(),
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
         ).getOrThrow()
-        val serializedCredential = credential.credentials.shouldNotBeEmpty().first().credentialString.shouldNotBeNull()
+            .shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Plain>()
+            .response
+        val serializedCredential = credential.credentials.shouldNotBeEmpty()
+            .first().credentialString.shouldNotBeNull()
 
         JwsSigned.deserialize<VerifiableCredentialJws>(
             VerifiableCredentialJws.serializer(),
@@ -197,7 +200,11 @@ val OidvciCodeFlowTest by testSuite {
                     clientNonce = issuer.nonceWithDpopNonce().getOrThrow().response.clientNonce,
                 ).getOrThrow().first(),
                 credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-            ).getOrThrow().credentials.shouldNotBeEmpty().first().credentialString.shouldNotBeNull()
+            ).getOrThrow()
+                .shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Plain>()
+                .response
+                .credentials.shouldNotBeEmpty().first()
+                .credentialString.shouldNotBeNull()
         }
     }
 
@@ -223,10 +230,12 @@ val OidvciCodeFlowTest by testSuite {
 
         val credentials: Collection<CredentialResponseSingleCredential> =
             issuer.credential(
-                token.toHttpHeaderValue(),
-                credentialRequest,
+                authorizationHeader = token.toHttpHeaderValue(),
+                params = WalletService.CredentialRequest.Plain(credentialRequest),
                 credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
             ).getOrThrow()
+                .shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Plain>()
+                .response
                 .credentials.shouldNotBeEmpty().shouldHaveSize(2)
         // subject identifies the key of the client, here the keys of different proofs, so they should be unique
         credentials.map {
@@ -275,7 +284,10 @@ val OidvciCodeFlowTest by testSuite {
             ).getOrThrow().first(),
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
         ).getOrThrow()
-        val serializedCredential = credential.credentials.shouldNotBeEmpty().first().credentialString.shouldNotBeNull()
+            .shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Plain>()
+            .response
+        val serializedCredential = credential.credentials.shouldNotBeEmpty()
+            .first().credentialString.shouldNotBeNull()
 
         serializedCredential.assertSdJwtReceived()
     }
@@ -296,7 +308,10 @@ val OidvciCodeFlowTest by testSuite {
             ).getOrThrow().first(),
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
         ).getOrThrow()
-        val serializedCredential = credential.credentials.shouldNotBeEmpty().first().credentialString.shouldNotBeNull()
+            .shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Plain>()
+            .response
+        val serializedCredential = credential.credentials.shouldNotBeEmpty().first()
+            .credentialString.shouldNotBeNull()
 
         serializedCredential.assertSdJwtReceived()
     }
@@ -351,7 +366,10 @@ val OidvciCodeFlowTest by testSuite {
             ).getOrThrow().first(),
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
         ).getOrThrow()
-        val serializedCredential = credential.credentials.shouldNotBeEmpty().first().credentialString.shouldNotBeNull()
+            .shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Plain>()
+            .response
+        val serializedCredential = credential.credentials.shouldNotBeEmpty()
+            .first().credentialString.shouldNotBeNull()
 
         serializedCredential.assertSdJwtReceived()
     }
@@ -378,7 +396,10 @@ val OidvciCodeFlowTest by testSuite {
             ).getOrThrow().first(),
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
         ).getOrThrow()
-        val serializedCredential = credential.credentials.shouldNotBeEmpty().first().credentialString.shouldNotBeNull()
+            .shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Plain>()
+            .response
+        val serializedCredential = credential.credentials.shouldNotBeEmpty()
+            .first().credentialString.shouldNotBeNull()
 
         serializedCredential.assertSdJwtReceived()
     }
@@ -511,7 +532,10 @@ val OidvciCodeFlowTest by testSuite {
             ).getOrThrow().first(),
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
         ).getOrThrow()
-        val serializedCredential = credential.credentials.shouldNotBeEmpty().first().credentialString.shouldNotBeNull()
+            .shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Plain>()
+            .response
+        val serializedCredential = credential.credentials.shouldNotBeEmpty()
+            .first().credentialString.shouldNotBeNull()
 
         val issuerSigned =
             coseCompliantSerializer.decodeFromByteArray<IssuerSigned>(serializedCredential.decodeToByteArray(Base64()))