+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Release 5.10.0 (Unreleased):
    - Update implementation to 1.0 from 2025-09-16
    - Remove code elements deprecated in 5.9.0
    - Drop single `proof` in credential request
+   - Support credential response encryption correctly, see changed API in `CredentialIssuer.credential()`
  - OpenID for Verifiable Presentations:
    - Update implementation to 1.0 from 2025-07-09
    - Remove code elements deprecated in 5.9.0